Everything still broken with scheduling! :(
Entity-Actor-Action ownership chain broken? Not sure what is going
wrong. When the actor tries to access it's owner during the update, it
should be accessing the Entity that owns that component, but it doesn't?
Possibly abandoning this idea for now, and moving back to the your-turn,
my-turn of the Python tutorial.