Changed entityList_ in Engine from a std::vector to a wsl::DLList
No real reason for this change, other than to see if it works. And it does! Not giving up on the scheduling system, yet. Tempted to shove all of the Entity logic into the Entity class, instead of separating it into components.
Design-wise, I know this is a bad idea - but it may work and may not be the worst thing.