Rough save/load finished!
Saving happens on quit, and loading happens when the game is opened.
Still has a few kinks to work out, but everything works and saves/loads
as it's supposed to. Saving/loading wsl::DLList and wsl::PQList required
a tricky workaround - which then meant that saving/loading objects that
owned one of those objects required a workaround. Basically, a new type
of Cereal object was created as a transition object for serialization.