Fixed PQList::remove()
The strange bug from changing the inventory over to a PQList was caused
by removing a node from the list. Since PQList is singly linked, yanking
a random node out caused a dangling pointer to a next node that didn't
exist. Simple fix, remove now looks for the previous node of the node to
be removed, then changes it's next to the removed node's next.