Began work on animations
Created rough classes - Basically an animation is a stack of frames,
each frame containing a list of tiles with a position and a duration.
After the duration has elapsed, the list moves on to the next frame
until the animation is completed. Functions will be written to return
animations of specific types, and Engine will store a stack of
animations. Engine will loop through it's stack of animations each
update, calling update on each animation in the stack and culling dead
(finished) animations from the stack. Draw will loop through the
animation stack after everything else has been rendered to the console,
but before SDL draws the console on the window.