Work on implementing animations
Well, the animation part of things is hella broken. Segfaults that I
can't pin down. The engine update/draw routine works fine, so the only
thing that needs to be reworked is the AnimationTile/Frame/Animation
classes. The game loop now updates at a nice fixed 60fps following the
Game Programming Patterns game loop example. Engine::update() accepts a
delta time variable - which is the amount of time since the last update.
Animation::update() will obviously be the main use for this, once
everything works. Committing as is now - but I plan to change ALL of the
code in animation.hpp/cpp shortly. Nothing works and I can't figure it
out, so I might as well start over from scratch!